Moves descriptor types out to their own files
